# swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "This is feedbacky swagger docs",
    "version": "1.0.0",
    "title": "Swagger Feedbacky",
    "termsOfService": "http://swagger.io/terms/",
    "contact": {
      "email": "contact@berkekurnaz.com"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
  "host": "feedbacky.swagger.io",
  "basePath": "/v2",
  "tags": [
    {
      "name": "User",
      "description": "All user operations"
    },
    {
      "name": "Form",
      "description": "All form operations"
    },
    {
      "name": "Answer",
      "description": "All answer operations"
    }
  ],
  "schemes": [
    "https",
    "http"
  ],
  "paths": {
    "/user/createUser": {
      "post": {
        "tags": [
          "User"
        ],
        "summary": "Create a new user",
        "description": "Create a new user",
        "operationId": "createUser",
        "produces": [
          "application/xml",
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Created user object",
            "required": true,
            "schema": {
              "$ref": "#/definitions/User"
            }
          }
        ],
        "responses": {
          "default": {
            "description": "successful operation"
          }
        }
      }
    },
    "/user/loginUser": {
      "post": {
        "tags": [
          "User"
        ],
        "summary": "Login user",
        "description": "",
        "operationId": "loginUser",
        "produces": [
          "application/xml",
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "List of user object",
            "required": true,
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/User"
              }
            }
          }
        ],
        "responses": {
          "default": {
            "description": "successful operation"
          }
        }
      }
    },
    "/form/createForm": {
      "post": {
        "tags": [
          "Form"
        ],
        "summary": "Create a new form",
        "description": "",
        "operationId": "createForm",
        "produces": [
          "application/xml",
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "List of user object",
            "required": true,
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/User"
              }
            }
          }
        ],
        "responses": {
          "default": {
            "description": "successful operation"
          }
        }
      }
    },
    "/form/getFormDetails": {
      "get": {
        "tags": [
          "Form"
        ],
        "summary": "Get Form Details",
        "description": "",
        "operationId": "getFormDetails",
        "produces": [
          "application/xml",
          "application/json"
        ],
        "parameters": [
          {
            "name": "username",
            "in": "query",
            "description": "The user name for login",
            "required": true,
            "type": "string"
          },
          {
            "name": "password",
            "in": "query",
            "description": "The password for login in clear text",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "type": "string"
            },
            "headers": {
              "X-Rate-Limit": {
                "type": "integer",
                "format": "int32",
                "description": "calls per hour allowed by the user"
              },
              "X-Expires-After": {
                "type": "string",
                "format": "date-time",
                "description": "date in UTC when token expires"
              }
            }
          },
          "400": {
            "description": "Invalid username/password supplied"
          }
        }
      }
    },
    "/form/getFormAnswers": {
      "get": {
        "tags": [
          "Form"
        ],
        "summary": "Get Form Answers",
        "description": "",
        "operationId": "getFormAnswers",
        "produces": [
          "application/xml",
          "application/json"
        ],
        "parameters": [],
        "responses": {
          "default": {
            "description": "successful operation"
          }
        }
      }
    },
    "/answer/createAnswer": {
      "post": {
        "tags": [
          "Answer"
        ],
        "summary": "Create a new answer",
        "description": "",
        "operationId": "createAnswer",
        "produces": [
          "application/xml",
          "application/json"
        ],
        "parameters": [],
        "responses": {
          "default": {
            "description": "successful operation"
          }
        }
      }
    }
  },
  "securityDefinitions": {
    "petstore_auth": {
      "type": "oauth2",
      "authorizationUrl": "http://petstore.swagger.io/oauth/dialog",
      "flow": "implicit",
      "scopes": {
        "write:pets": "modify pets in your account",
        "read:pets": "read your pets"
      }
    },
    "api_key": {
      "type": "apiKey",
      "name": "api_key",
      "in": "header"
    }
  },
  "definitions": {
    "User": {
      "type": "object",
      "properties": {
        "_id": {
          "type": "string"
        },
        "name": {
          "type": "string"
        },
        "mail": {
          "type": "string"
        },
        "password": {
          "type": "string"
        }
      },
      "xml": {
        "name": "User"
      }
    },
    "Form": {
      "type": "object",
      "properties": {
        "_id": {
          "type": "string"
        },
        "name": {
          "type": "string"
        },
        "fields": {
          "type": "string"
        },
        "userId": {
          "type": "string"
        }
      },
      "xml": {
        "name": "Tag"
      }
    },
    "Answer": {
      "type": "object",
      "required": [
        "name",
        "photoUrls"
      ],
      "properties": {
        "_id": {
          "type": "string"
        },
        "response": {
          "type": "string"
        },
        "userAgent": {
          "type": "string"
        },
        "formId": {
          "type": "string"
        }
      },
      "xml": {
        "name": "Pet"
      }
    },
    "ApiResponse": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "type": {
          "type": "string"
        },
        "message": {
          "type": "string"
        }
      }
    }
  },
  "externalDocs": {
    "description": "Find out more about Swagger",
    "url": "http://swagger.io"
  }
}